[WIP] Change managed objects to run asynchronously
The `read*`, `write*`, `create*` and `destroy*` methods of
the `SNMPv2-SMI` MIB objects redesigned to return immediately
and deliver their results via a call back. This change makes
it possible to work with the data that managed objects
represent through SNMP in an asynchronous manner.

This change breaks backward compatibility.